Topic1

+++ app/controllers/admin/foods_controller.rb
@@ -31,13 +31,17 @@ class Admin::FoodsController < ApplicationController
 
   def update
     @food = Food.find(params[:id])
-    @food.update(food_params)
-    # もし販売停止中であれば
-    if @food.sale_status == false
-      cart_food = CartFood.where(food_id: @food.id)
-      cart_food.destroy_all
+    if @food.update(food_params)
+      # もし販売停止中であれば
+      if @food.sale_status == false
+        cart_food = CartFood.where(food_id: @food.id)
+        cart_food.destroy_all
+      end
+      redirect_to admin_food_path(@food.id)
+    else
+      @genre = Genre.all
+      render "edit"
     end
-    redirect_to admin_food_path(@food.id)
   end
 
   private
